fix(test): resolve CI test failures for SubscriptionService
Fix two failing tests in SubscriptionServiceTest:

1. UpdatePaymentMethod_Works (422 error on both net472/net8.0):
   SubscriptionUpdatePaymentMethodParams.BodyContent() was wrapping the
   union-typed request body in an extra 'body' key, sending
   {"body":{"type":"new",...}} instead of {"type":"new",...}.
   The API expects the union content directly without wrapping.

2. PreviewChangePlan_Works (LineItem deserialization failure on net472):
   TaxRate properties used float (System.Single) which overflows when
   Prism mock server generates boundary values like -3.4e+38. Changed
   to double (System.Double) to match JSON's inherent double-precision
   representation and avoid cross-framework parsing differences.

+++ src/DodoPayments.Client.Tests/Models/Subscriptions/SubscriptionPreviewChangePlanResponseTest.cs
@@ -1400,7 +1400,7 @@ public class LineItemSubscriptionTest : TestBase
         string expectedDescription = "description";
         string expectedName = "name";
         int expectedTax = 0;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
 
         Assert.Equal(expectedID, model.ID);
         Assert.Equal(expectedCurrency, model.Currency);
@@ -1482,7 +1482,7 @@ public class LineItemSubscriptionTest : TestBase
         string expectedDescription = "description";
         string expectedName = "name";
         int expectedTax = 0;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
 
         Assert.Equal(expectedID, deserialized.ID);
         Assert.Equal(expectedCurrency, deserialized.Currency);
@@ -1725,7 +1725,7 @@ public class AddonTest : TestBase
         int expectedQuantity = 0;
         ApiEnum<string, TaxCategory> expectedTaxCategory = TaxCategory.DigitalProducts;
         bool expectedTaxInclusive = true;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
         ApiEnum<string, AddonType> expectedType = AddonType.Addon;
         int expectedUnitPrice = 0;
         string expectedDescription = "description";
@@ -1800,7 +1800,7 @@ public class AddonTest : TestBase
         int expectedQuantity = 0;
         ApiEnum<string, TaxCategory> expectedTaxCategory = TaxCategory.DigitalProducts;
         bool expectedTaxInclusive = true;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
         ApiEnum<string, AddonType> expectedType = AddonType.Addon;
         int expectedUnitPrice = 0;
         string expectedDescription = "description";
@@ -2045,7 +2045,7 @@ public class MeterTest : TestBase
         string expectedPricePerUnit = "price_per_unit";
         int expectedSubtotal = 0;
         bool expectedTaxInclusive = true;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
         ApiEnum<string, MeterType> expectedType = MeterType.Meter;
         string expectedUnitsConsumed = "units_consumed";
         string expectedDescription = "description";
@@ -2124,7 +2124,7 @@ public class MeterTest : TestBase
         string expectedPricePerUnit = "price_per_unit";
         int expectedSubtotal = 0;
         bool expectedTaxInclusive = true;
-        float expectedTaxRate = 0;
+        double expectedTaxRate = 0;
         ApiEnum<string, MeterType> expectedType = MeterType.Meter;
         string expectedUnitsConsumed = "units_consumed";
         string expectedDescription = "description";
